# Add wide/multiple char string and nerd font support for chunk